Fix clipping sprites when scrolling

+++ src/player.cpp
@@ -173,7 +173,7 @@ Player::update(InputMode input_mode, u8 pressed, u8 held) {
 }
 
 void Player::render(int y_scroll) {
-  u8 reference_y = (u8)(board.origin_y - y_scroll);
+  int reference_y = board.origin_y - y_scroll;
   static u8 animation_frame;
   static State current_state = State::Dead;
   CORO_RESET_WHEN(current_state != state);
